+++ source/consumers/add_ons/FilterBlock.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <consumers/add_ons/value_filter.h>
+#include <consumers/add_ons/ValueFilter.h>
 
 namespace DAQuiri {
