+++ tests/consumers/add_ons/ValueFilterTest.cpp
@@ -1,5 +1,5 @@
 #include <gtest/gtest.h>
-#include <consumers/add_ons/value_filter.h>
+#include <consumers/add_ons/ValueFilter.h>
 
 using namespace DAQuiri;
